mihomo: remove pre-download for lite release

+++ openwrt/scripts/10-customize-config.sh
@@ -3,17 +3,21 @@ rm -rf package/new/helloworld/luci-app-ssr-plus
 rm -rf package/new/helloworld/patch-luci-app-ssr-plus.patch
 
 # add mihomo
-git clone https://$github/pmkol/openwrt-mihomo package/new/openwrt-mihomo
 rm -rf package/new/helloworld/luci-app-mihomo
 rm -rf package/new/helloworld/mihomo
-if curl -s "https://$mirror/openwrt/23-config-common" | grep -q "^CONFIG_PACKAGE_luci-app-mihomo=y"; then
-    mkdir -p files/etc/mihomo/run/ui
-    curl -Lso files/etc/mihomo/run/geoip.metadb https://$github/MetaCubeX/meta-rules-dat/releases/download/latest/geoip-lite.metadb
-    curl -Lso files/etc/mihomo/run/ASN.mmdb https://$github/MetaCubeX/meta-rules-dat/releases/download/latest/GeoLite2-ASN.mmdb
-    curl -Lso metacubexd-gh-pages.tar.gz https://$github/MetaCubeX/metacubexd/archive/refs/heads/gh-pages.tar.gz
-    tar zxf metacubexd-gh-pages.tar.gz
-    rm metacubexd-gh-pages.tar.gz
-    mv metacubexd-gh-pages files/etc/mihomo/run/ui/metacubexd
+if [ "$MINIMAL_BUILD" = "y" ]; then
+    git clone https://$github/morytyann/OpenWrt-mihomo package/new/openwrt-mihomo
+else
+    git clone https://$github/pmkol/openwrt-mihomo package/new/openwrt-mihomo
+    if curl -s "https://$mirror/openwrt/23-config-common" | grep -q "^CONFIG_PACKAGE_luci-app-mihomo=y"; then
+        mkdir -p files/etc/mihomo/run/ui
+        curl -Lso files/etc/mihomo/run/geoip.metadb https://$github/MetaCubeX/meta-rules-dat/releases/download/latest/geoip-lite.metadb
+        curl -Lso files/etc/mihomo/run/ASN.mmdb https://$github/MetaCubeX/meta-rules-dat/releases/download/latest/GeoLite2-ASN.mmdb
+        curl -Lso metacubexd-gh-pages.tar.gz https://$github/MetaCubeX/metacubexd/archive/refs/heads/gh-pages.tar.gz
+        tar zxf metacubexd-gh-pages.tar.gz
+        rm metacubexd-gh-pages.tar.gz
+        mv metacubexd-gh-pages files/etc/mihomo/run/ui/metacubexd
+    fi
 fi
 
 # add ddns-go
